requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pdfplumber==0.11.0
weasyprint==68.1
jinja2==3.1.4
python-multipart==0.0.9
pytest==8.3.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
PyJWT>=2.8.0
httpx==0.27.0
python-docx==1.1.2
python-pptx==1.0.2
google-genai>=1.0.0
python-dotenv>=1.0.0
chromadb==0.6.3
